post

+++ app/Http/Controllers/TasksController.php
@@ -6,6 +6,8 @@ use Illuminate\Http\Request;
 
 use App\Task;    // 追加
 
+use App\Http\Controllers\Controller;
+
 class TasksController extends Controller
 {
     /**
@@ -16,10 +18,24 @@ class TasksController extends Controller
     public function index()
     {
         $tasks = Task::all();
-
-        return view('tasks.index', [
+            return view('tasks.index', [
             'tasks' => $tasks,
         ]);
+        
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('tasks.index', $data);
+        }else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -49,10 +65,15 @@ class TasksController extends Controller
             'content' => 'required|max:191',
         ]);
 
-        $task = new Task;
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
+        // $task = new Task;
+        // $task->status = $request->status;    // 追加
+        // $task->content = $request->content;
+        // $task->save();
+        
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
 
         return redirect('/');
     }
@@ -96,17 +117,23 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-            'status' => 'required|max:10',   // 追加
-            'content' => 'required|max:191',
-        ]);
+        $task = \App\Task::find($id);
         
-        $task = Task::find($id);
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
+        if (\Auth::id() === $task->user_id) {
+            $this->validate($request, [
+                'status' => 'required|max:10',   // 追加
+                'content' => 'required|max:191',
+            ]);
+        
+            $task = Task::find($id);
+            $task->status = $request->status;    // 追加
+            $task->content = $request->content;
+            $task->save();
 
-        return redirect('/');
+            return redirect('/');
+        }else {
+            return redirect('/');
+        } 
     }
 
     /**
@@ -117,9 +144,15 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
+        
+        if (\Auth::id() === $task->user_id) {
+            $task = Task::find($id);
+            $task->delete();
 
-        return redirect('/');
+            return redirect('/');
+        }else {
+            return redirect('/');
+        } 
     }
 }
