+++ resources/views/tasks/show.blade.php
@@ -2,6 +2,7 @@
 
 @section('content')
 
+<?php $user = $task->user; ?>
     <h1>id = {{ $task->id }} の詳細ページ</h1>
 
     <table class="table table-bordered">
@@ -18,11 +19,12 @@
             <td>{{ $task->content }}</td>
         </tr>
     </table>
-    
-    {!! link_to_route('tasks.edit', 'このタスクを編集', ['id' => $task->id], ['class' => 'btn btn-default']) !!}
-    
-    {!! Form::model($task, ['route' => ['tasks.destroy', $task->id], 'method' => 'delete']) !!}
-        {!! Form::submit('削除', ['class' => 'btn btn-danger']) !!}
-    {!! Form::close() !!}
+    @if (Auth::id() == $task->user_id)
+        {!! link_to_route('tasks.edit', 'このタスクを編集', ['id' => $task->id], ['class' => 'btn btn-default']) !!}
+        
+        {!! Form::model($task, ['route' => ['tasks.destroy', $task->id], 'method' => 'delete']) !!}
+            {!! Form::submit('削除', ['class' => 'btn btn-danger']) !!}
+        {!! Form::close() !!}
+    @endif
 
 @endsection
